Extract validate_schemas to s3._read
